+++ mlblocks/__init__.py
@@ -5,3 +5,11 @@
 __author__ = 'MIT Data To AI Lab'
 __email__ = 'dailabmit@gmail.com',
 __version__ = '0.1.8-dev'
+
+import importlib
+
+
+def import_object(object_name):
+    """Import an object from its Fully Qualified Name."""
+    package, name = object_name.rsplit('.', 1)
+    return getattr(importlib.import_module(package), name)
