--- mlblocks/parsers/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-
-import importlib
-
-
-def import_object(object_name):
-    """Import an object from its Fully Qualified Name."""
-    package, name = object_name.rsplit('.', 1)
-    return getattr(importlib.import_module(package), name)
