+++ mlblocks/parsers/json.py
@@ -1,6 +1,6 @@
 from mlblocks.mlblock import MLBlock
 from mlblocks.mlhyperparam import MLHyperparam
-from mlblocks.parsers import import_object
+from mlblocks import import_object
 
 
 class MLJsonParser(object):
